Code re-org and re-use
