+++ africanus/averaging/shared.py
@@ -1,26 +1,152 @@
 # -*- coding: utf-8 -*-
 
-from africanus.util.numba import generated_jit, njit
+import numpy as np
 
-@njit(nogil=True, inline='always')
+from africanus.util.numba import is_numba_type_none, generated_jit, njit
+
+
+@generated_jit(nopython=True, nogil=True, cache=True)
+def shape_or_invalid_shape(array, ndim):
+    """ Return array shape tuple or (-1,)*ndim if the array is None """
+
+    try:
+        ndim_lit = getattr(ndim, "literal_value")
+    except AttributeError:
+        raise ValueError("ndim must be a integer literal")
+
+    if is_numba_type_none(array):
+        tup = (-1,)*ndim_lit
+
+        def impl(array, ndim):
+            return tup
+    else:
+        def impl(array, ndim):
+            return array.shape
+
+    return impl
+
+
+# TODO(sjperkins)
+# maybe inline='always' if
+# https://github.com/numba/numba/issues/4693 is resolved
+@njit(nogil=True, cache=True)
+def find_chan_corr(chan, corr, shape, chan_idx, corr_idx):
+    """
+    1. Get channel and correlation from shape if not set and the shape is valid
+    2. Check they agree if they already agree
+
+    Parameters
+    ----------
+    chan : int
+        Existing channel size
+    corr : int
+        Existing correlation size
+    shape : tuple
+        Array shape tuple
+    chan_idx : int
+        Index of channel dimension in ``shape``.
+    corr_idx : int
+        Index of correlation dimension in ``shape``.
+
+    Returns
+    -------
+    int
+        Modified channel size
+    int
+        Modified correlation size
+    """
+    if chan_idx != -1:
+        array_chan = shape[chan_idx]
+
+        # Corresponds to a None array, ignore
+        if array_chan == -1:
+            pass
+        # chan is not yet set, assign
+        elif chan == 0:
+            chan = array_chan
+        # Check consistency
+        elif chan != array_chan:
+            raise ValueError("Inconsistent Channel Dimension "
+                             "in Input Arrays")
+
+    if corr_idx != -1:
+        array_corr = shape[corr_idx]
+
+        # Corresponds to a None array, ignore
+        if array_corr == -1:
+            pass
+        # corr is not yet set, assign
+        elif corr == 0:
+            corr = array_corr
+        # Check consistency
+        elif corr != array_corr:
+            raise ValueError("Inconsistent Correlation Dimension "
+                             "in Input Arrays")
+
+    return chan, corr
+
+
+# TODO(sjperkins)
+# maybe inline='always' if
+# https://github.com/numba/numba/issues/4693 is resolved
+@njit(nogil=True, cache=True)
+def chan_corrs(vis, flag,
+               weight_spectrum, sigma_spectrum,
+               chan_freq, chan_width,
+               effective_bw, resolution):
+    """
+    Infer channel and correlation size from input dimensions
+
+    Returns
+    -------
+    int
+        channel size
+    int
+        correlation size
+    """
+    vis_shape = shape_or_invalid_shape(vis, 3)
+    flag_shape = shape_or_invalid_shape(flag, 3)
+    weight_spectrum_shape = shape_or_invalid_shape(weight_spectrum, 3)
+    sigma_spectrum_shape = shape_or_invalid_shape(sigma_spectrum, 3)
+    chan_freq_shape = shape_or_invalid_shape(chan_freq, 1)
+    chan_width_shape = shape_or_invalid_shape(chan_width, 1)
+    effective_bw_shape = shape_or_invalid_shape(effective_bw, 1)
+    resolution_shape = shape_or_invalid_shape(resolution, 1)
+
+    chan = 0
+    corr = 0
+
+    chan, corr = find_chan_corr(chan, corr, vis_shape, 1, 2)
+    chan, corr = find_chan_corr(chan, corr, flag_shape, 1, 2)
+    chan, corr = find_chan_corr(chan, corr, weight_spectrum_shape, 1, 2)
+    chan, corr = find_chan_corr(chan, corr, sigma_spectrum_shape, 1, 2)
+    chan, corr = find_chan_corr(chan, corr, chan_freq_shape, 0, -1)
+    chan, corr = find_chan_corr(chan, corr, chan_width_shape, 0, -1)
+    chan, corr = find_chan_corr(chan, corr, effective_bw_shape, 0, -1)
+    chan, corr = find_chan_corr(chan, corr, resolution_shape, 0, -1)
+
+    return chan, corr
+
+
+@njit(nogil=True)
 def flags_match(flag_row, ri, out_flag_row, ro):
     if flag_row is None:
         return True
     else:
         return flag_row[ri] == out_flag_row[ro]
 
-@njit(nogil=True, inline='always')
+@njit(nogil=True)
 def is_chan_flagged(flag, r, f, c):
     return False if flag is None else flag[r, f, c]
 
 
-@njit(nogil=True, inline='always')
-def chan_adder(output, input, orow, ochan, irow, ichan, corr):
+@njit(nogil=True)
+def chan_add(output, input, orow, ochan, irow, ichan, corr):
     if input is not None:
         output[orow, ochan, corr] += input[irow, ichan, corr]
 
 
-@njit(nogil=True, inline='always')
+@njit(nogil=True)
 def vis_add(out_vis, out_weight_sum, in_vis,
             weight, weight_spectrum,
             orow, ochan, irow, ichan, corr):
@@ -46,7 +172,7 @@ def vis_add(out_vis, out_weight_sum, in_vis,
         out_weight_sum[orow, ochan, corr] += 1.0
 
 
-@njit(nogil=True, inline='always')
+@njit(nogil=True)
 def sigma_spectrum_add(out_sigma, out_weight_sum, in_sigma,
                        weight, weight_spectrum,
                        orow, ochan, irow, ichan, corr):
@@ -75,7 +201,7 @@ def sigma_spectrum_add(out_sigma, out_weight_sum, in_sigma,
 
 
 
-@njit(nogil=True, inline='always')
+@njit(nogil=True)
 def normalise_vis(vis_out, vis_in, row, chan, corr, weight_sum):
     if vis_in is not None:
         wsum = weight_sum[row, chan, corr]
@@ -84,9 +210,9 @@ def normalise_vis(vis_out, vis_in, row, chan, corr, weight_sum):
             vis_out[row, chan, corr] = vis_in[row, chan, corr] / wsum
 
 
-@njit(nogil=True, inline='always')
+@njit(nogil=True)
 def normalise_sigma_spectrum(sigma_out, sigma_in, row, chan, corr, weight_sum):
-    if sigma_in is not None:
+    if sigma_in is not None and weight_sum is not None:
         wsum = weight_sum[row, chan, corr]
 
         if wsum == 0.0:
@@ -96,7 +222,48 @@ def normalise_sigma_spectrum(sigma_out, sigma_in, row, chan, corr, weight_sum):
         res = np.sqrt(sigma_in[row, chan, corr] / (wsum**2))
         sigma_out[row, chan, corr] = res
 
-@njit(nogil=True, inline='always')
+
+# @generated_jit(nopython=True, nogil=True)
+# def normalise_sigma_spectrum(sigma_out, sigma_in, row, chan, corr, weight_sum):
+#     def impl(sigma_out, sigma_in, row, chan, corr, weight_sum):
+#         if sigma_in is not None and weight_sum is not None:
+#             wsum = weight_sum[row, chan, corr]
+
+#             if wsum == 0.0:
+#                 return
+
+#             # sqrt(sigma**2 * weight**2 / (weight(sum**2)))
+#             res = np.sqrt(sigma_in[row, chan, corr] / (wsum**2))
+#             sigma_out[row, chan, corr] = res
+
+#     return impl
+
+
+# @generated_jit(nopython=True, nogil=True)
+# def normalise_weight_spectrum(wt_spec_out, wt_spec_in, row, chan, corr):
+#     def impl(wt_spec_out, wt_spec_in, row, chan, corr):
+#         if wt_spec_in is not None:
+#             wt_spec_out[row, chan, corr] = wt_spec_in[row, chan, corr]
+
+#     return impl
+
+
+@njit(nogil=True)
+def normalise_sigma_spectrum(sigma_out, sigma_in, row, chan, corr, weight_sum):
+    if sigma_in is not None and weight_sum is not None:
+        wsum = weight_sum[row, chan, corr]
+
+        if wsum == 0.0:
+            return
+
+        # sqrt(sigma**2 * weight**2 / (weight(sum**2)))
+        res = np.sqrt(sigma_in[row, chan, corr] / (wsum**2))
+        sigma_out[row, chan, corr] = res
+
+
+@njit(nogil=True)
 def normalise_weight_spectrum(wt_spec_out, wt_spec_in, row, chan, corr):
     if wt_spec_in is not None:
         wt_spec_out[row, chan, corr] = wt_spec_in[row, chan, corr]
+
+
